+++ CHANGELOG.md
@@ -96,6 +96,8 @@
 - `pcs cluster cib-upgrade` no longer exits with an error if the CIB schema is
   already the latest available (this has been changed in pacemaker-2.0)
 - Pcs now configures corosync to put timestamps in its log ([rhbz#1615420])
+- Option `-V` has been replaced with `--full` and a CIB file can be specified
+  only using option `-f` in `pcs cluster verify`
 
 ### Security
 - CVE-2018-1086: Debug parameter removal bypass, allowing information disclosure